refactoring

+++ src/main/java/github/dqw4w9wgxcq/farmproxyservice/service/awscheckip/AwsCheckIpResult.java
@@ -1,4 +1,4 @@
 package github.dqw4w9wgxcq.farmproxyservice.service.awscheckip;
 
-public record AwsCheckIpResult(String ip, long latency) {
+public record AwsCheckIpResult(String ip, int latency) {
 }
